requirements.txt
chromadb==0.4.18
google-cloud-aiplatform==1.37.0
langchain==0.0.138
oauthlib==3.2.2
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
sentence-transformers==2.2.2
streamlit==1.24.1
transformers==4.29.0
urllib3==1.26.14
feedparser==6.0.10
